Add basic github actions CI workflow
This tests whether the package compiles and imports successfully.

I also added Numpy to setup_requires due to the f2py compilation step
requiring it.

This is not called with `python setup.py install/develop`, and has to be
done manually, so I added a step that install requirements beforehand.

I also added `pexpect` for omfit_commonclasses, which seemed to fail on
import for lack of it; this should be removed when possible and
double-checked with the great folks at OMFIT - seems like a missing
requirement :)

Fail-fast in the `strategy` part allows all the builds to proceed
independently. See https://github.community/t/build-got-was-canceled-unexpectedly/16130